Chat, AllConversationsWindow are separated in view, viewmodel layers
