+++ ISSLab/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:ISSLab">
-    <!--StartupUri="MainWindow.xaml">-->
+    <!--StartupUri="View/AllConversationsWindow.xaml">-->
     <Application.Resources>
         
     </Application.Resources>
