+++ ISSLab/View/Chat.xaml.cs
@@ -35,46 +35,13 @@ namespace ISSLab.View
 
         public void SendMessage(string message, bool isMine, bool isSellingPost)
         {
-            var newMessage = new Message
-            {
-                Content = message,
-                Width = CalculateMessageWidth(message),
-                IsMine = isMine,
-                BubbleColor = isMine ? Brushes.LightBlue : Brushes.LightGray,
-                HorizontalAlignment = isMine ? HorizontalAlignment.Right : HorizontalAlignment.Left
-            };
-
-            if (isSellingPost)
-            {
-                newMessage.Content = "SELLING POST: " + "";
-                newMessage.BubbleColor = Brushes.YellowGreen;
-            }
-
-            _viewModel.AddMessage(newMessage);
-
+            _viewModel.SendMessage(message, isMine, isSellingPost);
             MessageTextBox.Text = "";
         }
+
         public void SendBuyingMessage(string media)
         {
-            var newMessage = new Message
-            {
-                Content = "I'm interested in buying your product!",
-                Width = CalculateMessageWidth("I'm interested in buying your product!"),
-                IsMine = false,
-                BubbleColor = Brushes.YellowGreen,
-                HorizontalAlignment = HorizontalAlignment.Left,
-                ImagePath = media,
-                ShowAcceptButton = true, 
-                ShowRejectButton = true
-            };
-            _viewModel.AddMessage(newMessage);
-        }
-
-        private double CalculateMessageWidth(string message)
-        {
-            var textBlock = new TextBlock { Text = message, TextWrapping = TextWrapping.Wrap };
-            textBlock.Measure(new Size(double.PositiveInfinity, double.PositiveInfinity));
-            return textBlock.DesiredSize.Width;
+            _viewModel.SendBuyingMessage(media);
         }
 
         private void SendButton_Click(object sender, RoutedEventArgs e)
