+++ ISSLab/ViewModel/ChatViewModel.cs
@@ -4,7 +4,10 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Media;
+using System.Windows;
 using ISSLab.Model;
+using System.Windows.Controls;
 
 namespace ISSLab.ViewModel
 {
@@ -23,6 +26,49 @@ namespace ISSLab.ViewModel
         public void AddMessage(Message message)
         {
             AllMessages.Add(message);
+        }
+
+        public void SendMessage(string message, bool isMine, bool isSellingPost)
+        {
+            var newMessage = new Message
+            {
+                Content = message,
+                Width = CalculateMessageWidth(message),
+                IsMine = isMine,
+                BubbleColor = isMine ? Brushes.LightBlue : Brushes.LightGray,
+                HorizontalAlignment = isMine ? HorizontalAlignment.Right : HorizontalAlignment.Left
+            };
+
+            if (isSellingPost)
+            {
+                newMessage.Content = "SELLING POST: " + "";
+                newMessage.BubbleColor = Brushes.YellowGreen;
+            }
+
+            AllMessages.Add(newMessage);
+        }
+
+        public void SendBuyingMessage(string media)
+        {
+            var newMessage = new Message
+            {
+                Content = "I'm interested in buying your product!",
+                Width = CalculateMessageWidth("I'm interested in buying your product!"),
+                IsMine = false,
+                BubbleColor = Brushes.YellowGreen,
+                HorizontalAlignment = HorizontalAlignment.Left,
+                ImagePath = media,
+                ShowAcceptButton = true,
+                ShowRejectButton = true
+            };
+            AllMessages.Add(newMessage);
+        }
+
+        private double CalculateMessageWidth(string message)
+        {
+            var textBlock = new TextBlock { Text = message, TextWrapping = TextWrapping.Wrap };
+            textBlock.Measure(new Size(double.PositiveInfinity, double.PositiveInfinity));
+            return textBlock.DesiredSize.Width;
         }
     }
 }
